feat: move product-related types to types/products.ts and update imports in slice and operations
